Force Powershop RSpec mode to start a persistent spring server if it's not running
For some reason the in-built spring feature of rspec-mode starts a
spring server, but it dies immediately following the conclusion of the
initial RSpec compilation.

This change manually starts a spring server for the requested
Powershop market (one that actually persists after the initial RSpec
compilation). The code is ugly AF, but it works ¯\_(ツ)_/¯.